[#55] fix issue
