+++ app/modules/form-place/components/add/form-place-add.component.ts
@@ -1,22 +1,27 @@
 import { Component, OnInit } from "@angular/core";
 import { Router } from "@angular/router";
 
-import { PlaceSearchService } from '../../../../shared/place/place-search.service';
+import { LoadingIndicator } from "nativescript-loading-indicator";
+
+import { PlaceSearchService } from '../../../../shared/place/search/place-search.service';
+import providePlaceStorageService from '../../../../shared/place/storage/provide-service';
+import { PlaceStorageService } from '../../../../shared/place/storage/place-storage.service';
 
 @Component({
   moduleId: module.id,
   selector: 'kl-form-place-add',
   templateUrl: 'form-place-add.html',
+  providers: [ providePlaceStorageService() ]
 })
 export class FormPlaceAddComponent implements OnInit{
   protected step:string = 'search';
   protected place:any;
 
-  public constructor( private placeSearch: PlaceSearchService, private router:Router) { }
+  public constructor( private placeSearch: PlaceSearchService, private placeStorage: PlaceStorageService, private router:Router) { }
 
   public ngOnInit() { }
 
-  protected onSearchPlace(place) {
+  protected onFindPlace(place) {
     this.placeSearch.searchById(place.id)
       .subscribe(
         (place) => {
@@ -24,8 +29,19 @@ export class FormPlaceAddComponent implements OnInit{
           this.place.contexts = [];
           this.step = 'validation';
         }
-      )
+      );
+  }
 
-    //this.router.navigate(['/add-place-validation', place.id, 'geoloc']);
+  public onSubmit() {
+    const loader = new LoadingIndicator();
+    loader.show({
+      message: 'process ...',
+      ios: {
+        color: '#FFFFFF',
+        backgroundColor: '#000000',
+      }
+    });
+
+    this.placeStorage.insert(this.place).subscribe((place => { this.place = place; loader.hide() }));
   }
 }
