+++ app/modules/form-place/components/add/form-place-add.html
@@ -1,4 +1,4 @@
 <StackLayout orientation="vertical">
-  <kl-form-place-search (submit)="onSearchPlace($event)" *ngIf="step === 'search'"></kl-form-place-search>
-  <kl-form-place-validation *ngIf="step === 'validation'" [place]="place"></kl-form-place-validation>
+  <kl-form-place-search (submit)="onFindPlace($event)" *ngIf="step === 'search'"></kl-form-place-search>
+  <kl-form-place-validation *ngIf="step === 'validation'" [place]="place" (submit)="onSubmit($event)"></kl-form-place-validation>
 </StackLayout>
