+++ app/modules/form-place/components/edit/form-place-edit.html
@@ -1,3 +1,3 @@
 <StackLayout orientation="vertical">
-  <kl-form-place-validation [place]="place"></kl-form-place-validation>
+  <kl-form-place-validation [place]="place" (submit)="onSubmit($event)"></kl-form-place-validation>
 </StackLayout>
