+++ app/shared/place/storage/provide-service.ts
@@ -3,5 +3,5 @@ import { PlaceStorageMockService } from './place-storage.mock.service';
 import { Config } from '../../config';
 
 export default () => {
-  return { provide: PlaceStorageService, useFactory: () => Config.mock ? new PlaceStorageMockService() : new PlaceStorageService(), deps: [] }
+  return { provide: PlaceStorageService, useFactory: () => Config.mockStorage ? new PlaceStorageMockService() : new PlaceStorageService(), deps: [] }
 }
